use class syntax for Sekshi and always enable System module
