+++ package.json
@@ -1,8 +1,8 @@
 {
   "name": "sekshibot",
   "description": "Modular plug.dj room bot for WE ♥ KPOP",
-  "version": "0.9.1",
-  "main": "lib/sekshi",
+  "version": "0.10.0",
+  "main": "lib/Sekshi",
   "author": "Sooyou",
   "contributors": [
     "brookiebeast",
